EBIP-11 Cyfrin Changes

+++ protocol/contracts/ecosystem/oracles/UsdOracle.sol
@@ -20,7 +20,7 @@ contract UsdOracle {
         return LibEthUsdOracle.getEthUsdPrice();
     }
 
-    function getEthUsdTwa(uint32 lookback) external view returns (uint256) {
+    function getEthUsdTwa(uint256 lookback) external view returns (uint256) {
         return LibEthUsdOracle.getEthUsdPrice(lookback);
     }
