+++ protocol/contracts/libraries/Oracle/LibUsdOracle.sol
@@ -25,9 +25,12 @@ library LibUsdOracle {
 
     /**
      * @dev Returns the price of a given token in in USD with the option of using a lookback.
-     * `lookback` should be 0 if the instantaneous price is desired. Otherwise, the TWAP lookback in seconds.
+     * `lookback` should be 0 if the instantaneous price is desired. Otherwise, it should be the
+     * TWAP lookback in seconds.
+     * If using a non-zero lookback, it is recommended to use a substantially large `lookback`
+     * (> 900 seconds) to protect against manipulation.
      */
-    function getUsdPrice(address token, uint32 lookback) internal view returns (uint256) {
+    function getUsdPrice(address token, uint256 lookback) internal view returns (uint256) {
         if (token == C.WETH) {
             uint256 ethUsdPrice = LibEthUsdOracle.getEthUsdPrice(lookback);
             if (ethUsdPrice == 0) return 0;
